samples/rust: add simple mutex rust sample
It has a shared counter that gets incremented 100M times (each increment
involves locking, incrementing by 1, unlocking) when the device is
written to. When the device is read, it returns (as a string) the
current count.

This allows us to test the mutex implementation by doing several
simultaneous writes from the shell, and at the end, do a single read and
check that the number is 100M times the number of writes. If the lock is
allowing more than one thread to run in the critical region, the counter
will very likely be less than that.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>